+++ ngz/projects/ngz-common/src/lib/services/api/ApiEndpoint/adapters/index.ts
@@ -2,6 +2,7 @@
 // ----------------------------------------------------------------------------
 
 // (Re)export adapters(s)
+export * from './ApiEndpointBaseAdapter';
 export * from './ApiEndpointCrudComponentAdapter';
 export * from './ApiEndpointToGridAdapter';
 export * from './ApiEndpointToAutocompleteAdapter';